‘Moved separator under menu-folder and removed it from index
