+++ packages/hsl-shared-components/src/index.js
@@ -1,11 +1,6 @@
 import AppHeader from './AppHeader/AppHeader';
 import Button, { RoundButton, ButtonWithText } from './Button';
 import Input from './Input/Input';
-import Separator, {
-  MenuSeparator,
-  HorizontalSeparator,
-  HorizontalMenuSeparator
-} from './Separator/Separator';
 import Nav, { NavItem, NavMobile, NavDesktop } from './Nav';
 import PageTitleBar from './PageTitleBar/PageTitleBar';
 import Container from './Container/Container';
@@ -37,10 +32,6 @@ export {
   RoundButton,
   ButtonWithText,
   Input,
-  Separator,
-  MenuSeparator,
-  HorizontalSeparator,
-  HorizontalMenuSeparator,
   Nav,
   NavItem,
   NavMobile,
